chore(rename): Renames the font folder to be fonts.
This is in-keeping with the new convention about folder naming: folders should either be the extension of the files contained (if they're all the same), or the unabbreviated plural form of the noun of the kind of files contained - i.e. images, not img, and fonts, not font (similarly, directives, and templates, or html).